fix: get serial input from RFID reader and clean up Serial1 usage
Cleaned up some of the Serial.print statements so that Serial(115200) is used exclusively for serial debug output and Serial1(9600) is used exclusively for serial input from the RFID reader. It's unclear to me why prior authors mixed and matched printing debug statements on Serial and Serial1. Serial1 handling and card reading has been separated into a separate library, and this commit assumes the presence of a microswitch not present in the parent hardware. This commit marks the first time our assembled prototype achieved an RFID card read and subsequently confirmed that read via serial debugging.